BQ_remote_Ver2/requirements.txt
google-adk>=1.23.0

google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

mcp>=1.0.0

openpyxl>=3.1.0
